Create Matrix
Drill down is a feature in Power BI that allows users to go from a broad view of data to a more detailed one with a single mouse click.
Turn on the drill-down option by selecting the single downward arrow. 
The grey background indicates that the drill-down option is turned on. 
Another way to access the drill features is by right-clicking a data point on the visual to open a menu with available options.
We've got the "plus" icon next to the start of a hierarchy.
So we have a plus icon next to the year in this example and this allows us to drill down and drill up while still begin able to see all of the rest datas