Create README.md
In Power BI, we have many types of visualization techniques. The most important and extensively used one of them is charts. 
Charts are used in every report or dashboard to display data to a user. 
We have built-in Power BI charts to use but can also make combo charts or custom-made charts.

Here I have used Stacked Column Charts These are the basic typed charts that allow the comparison of one category to another category.
Stacked charts are a sort of bar chart which are multiple-bar in nature where related values are placed at one another. 
The main parts of Stacked Column Charts are:Title: It denotes the information about the chart.
X-axis: It is the individual entry for the category to be presented.
Y-axis: It is for the value against each type of category.
Legend:  It is the different category that will contribute to the charts.
Bars: These heights represent the total value of all the legends.


Creation of Stacked Column Charts
Step 1: Import data from Your Excel to Power BI. Go to Home Tab-> Get Data -> Choose Data which you want Example Excel -> Select the file and Open -> Select The sheet and Load.
Step 2: Load the column chart under visualization and click on the ‘stacked column chart’ icon.
Step 3: The chart will be loaded on the screen and resized if required.
Step 4: To start creating a column chart click on the fields which is to be used in a column chart.
Step 5: Drag and drops the fields that are to be placed on the X-axis and Y-axis of the chart respectively.
Step 6: A simple Column chart has been created as shown below.
Step 7: To show the category in different colors use ‘Legend’, and drag and drop the category which is to be shown in color.