Create README.md
Open Power BI and select the "Visualizations" pane.
Select the "Line chart" option from the available chart types.
Drag and drop the data you want to visualize into the "Values" field of the "Visualizations" pane.
Select the fields you want to display in your line chart, including the X-axis and Y-axis values as well as any Legend fields.
Drag and drop the two-column fields under the Y-axis and Secondary Y-axis section in the visualizations pane.

A Line chart shows data in the form of a line. It is generally used when we have to show, continuous data graphically. 
We could add a secondary y-axis, with just 2 mouse clicks in Power BI. 
We have various options to format line charts, we can change the value of the x-axis, y-axis, its title, etc.

Here I have taken sum of sales volume based  on the year and region name.Each dots in the line reprsesent the years.