Create README.md
In Power BI, a combo chart combines a line chart and a column chart into one visual. 
This allows you to quickly compare two sets of measures.
In this example we can see the individual regions and from that we are able to find who is above and below.
Here in my example West Midland not joining in with the house price influcation at this stage.
We can format the line section which is used to be in shape section.
And at the bottom we have got this shade area. So that it turns from a line and clustered column chart to an area and clustered column chart.