Create README.md
A Scatter chart is a chart, representing data points, spread on a graph, to show the column values. 
It helps us analyze and summarise the data very efficiently. We can also add legends, bubble sizes, and animations to this chart. 
This graph is widely used because it can show better visualizations in the case of both categorical as well as numeric data. 
In this article, we will learn how to create a scatter chart in Power BI.
In this project I am comparing sales volume and house price inflation .
One or the other way of highlighting a particular series is by clicking on it and you'll see that a trail happenes you can see where it has been and this also quite useful when you plat it.
You can see how it progress over the year.