Create README.md
Cards in Power BI help us display a certain value, corresponding to different dynamic scenarios. 
For example, want to know the total population of the world, and the population of a certain country by a slicer, all can be achieved by cards in Power BI. 
In this article, we will learn how to format a card in Power BI.
  
  Callout Value
Callout value helps format the main content of the card. 

  General Formatting 
There are multiple options in general formatting. For cards, we have options like Title, tooltip, effects, alt text, etc.
  Multi-Row cards in Power BI help us display multiple values, corresponding to different dynamic scenarios.
  Under the fields section, the first dropped column acts as the rows, and the further addition of columns, under the main column, refers to the columns of that card.
    After the successful, creation of a multi-row card in Power BI. We have multiple options to format it.
       Using this card, we will explore every option of the multi-row card in Power BI. There are two types of Formatting in visualizations i.e. visual formatting and general formatting.
      Visual formatting comprises three options, Callout value, and Category label, and Cards.Callout value helps format the main content of the card.
        Category label help customize the text written below the main text. For example, Sum of Salary and Sum of Bonus.
          The Cards option is only present in multi-row cards and is not present in single-element cards.
          There are multiple options in general formatting. For cards, we have options like Title, tooltip, effects, alt text, etc.