Create README.md
A decomposition tree in Power BI is a visual tool that organizes data into easy-to-understand formats. 
It helps analysts quickly identify patterns, correlations, and anomalies in datasets.

The visualization requires two types of input:

Analyze: This is the metric needed for analyzing. It must be a measure or an aggregate. This is the column the user needs for the breakdown of the data tree.
Explain By: This holds one or more dimensions needed to drill down into. With the help of these columns, further analysis can be done.

AI split helps in finding high and low data automatically.

High value: This shows the highest value of a field for any selected metric.
low value: This shows the lowest value of a field for any selected metric.

Benefits of Decomposition Tree:

It supports easy drill-down features.
It supports conditional formatting of nodes with different colors and custom changes.
It supports sort capabilities and cross-highlight features.
It supports improved filtering and keyboard shortcuts.

Decomposition trees analyze one value by one or multiple categories, or dimensions. 
Keep selecting high value until the user has a decomposition tree. 
You can delete levels by selecting the X in the heading whichever column is unwanted in that particular time. 
While trying out multiple dimensions in the decomposition tree or exploring the data, 
one can find the hierarchy and dataset of interest using the drill-down options.