Create README.md
A decomposition tree in Power BI visualizes data in its components or building blocks. It allows you to: 
Disaggregate data to understand what makes up a value
Automatically aggregate data
Drill down into dimensions in any order
Find the next dimension data based on defined criteria
Determine the high and low values using AI functions

We can add drill-through fields by dragging and dropping them in the bottom-most area in the drill-through section. 
Let’s say we want to drill through the data shown in the decomposition tree by an attribute named Brand. 
When we drag and drop this attribute in the Drill Through section, we would be able to see the distinct values in this field. 
If we select one of the values in this field as shown below, the data would be scoped to the selected value as shown below.
While exploring the data and trying out different measures and dimensions in the decomposition tree, 
one may eventually find the hierarchy and dataset of interest using the drill-down approach and drill-through options. 
Saving and publishing the report is one way of preserving the analysis. 
Another option one may want to exercise is to export the data in a tabular format, 
so that it can be used elsewhere outside of the report as well. 
This can be easily accomplished in Power BI by clicking on the top-right corner of the report and exporting the data in the decomposition tree as shown below.

We can use the top and down arrows shown at each level of the hierarchy to scroll through the data. 
While this remains an option, one would typically want to sort the data in an ascending or descending order, or even by a different attribute. 
We can accomplish the same as well by using the sort options provided in the context menu of the visualization.