Create README.md
You can append two or more tables. 
 Appending means that the results of two (or more) queries (which are tables themselves) will be combined into one query. 
 Rows will be appended after each other.
 The append operation creates a single table by adding the contents of one or more tables to another, 
 and aggregates the column headers from the tables to create the schema for the new table.
 Append operations join two or more tables vertically. 
The data rows from one table are appended (or added) at the end of the data rows in another table where the column values match. 
So in an append operation, the base table will have the same number of columns at the end of the operations as it did at the start, but each column will contain more rows.
Appending power queries in Power BI is an easy yet powerful technique in terms of setting up and optimizing models. 
We don’t need to create any new calculations, new filters, or new visualizations because it should all filter into anything we have built before. 
Combining this with the other techniques makes our model scalable from an analytical solution to a reporting perspective.