Create README.md
DAX functions play an important role in the usage of DAX for data modeling and reporting. 
It is an inbuilt function provided in the DAX language that helps you perform commonly used data calculations on the Data Model. 
In this tutorial, you will learn data modeling and analysis using DAX, based on the Power Pivot data model. 
As you proceed, you will learn about the different facets of Power Pivot, DAX and DAX functions in data modeling and analysis. 
By the end of the tutorial, you will be able to perform data modeling and analysis with DAX for any context at hand.
DAX stands for Data Analysis Expressions. 
DAX is a formula language and is a collection of functions, operators, and constants that can be used in a formula or expression to calculate and return one or more values. 
DAX is the formula language associated with the Data Model of Microsoft Excel Power Pivot and with Microsoft Power BI.
DAX is not a programming language, however it is a formula language that allows the users to define custom calculations in calculated columns and calculated fields (also known as measures). 
DAX helps you create new information from the existing data in your Data Model. 
DAX formulas enable you to perform data modeling, data analysis, and use the results for reporting and decision making.
A DAX function is an inbuilt function provided in the DAX language to enable you to perform various actions on the data in the tables in your Data Model.
DAX functions enable you to perform commonly used data calculations on the Data Model. 
Some of the DAX functions have same names and functionality as that of Excel functions but have been modified to use DAX data types and to work with tables and columns, 
as highlighted in the next section. DAX has additional functions that are designed to work with relational data and perform dynamic aggregation.
DAX functions play an important role in the usage of DAX for data modeling and reporting.