Create README.md
Data Model example 
Data modeling in Power BI is a topic that has been brought more to the forefront for even the introductory level courses. 
Understanding data modeling is crucial for improving your reports and outputs and saving you hours of frustration 
looking for the root cause of an error that may just be due to a wrong relationship in your data model.

In general terms, a data model in Power BI logically represents the relationships between two or more disparate physical 
tables connected to each other using unique identifiers, also known as Primary Keys and Foreign Keys. 
These sensible relationship structures could be constructed; however the users may choose to design it, which is known as the schema.