Create README.md
The AVERAGEX DAX function allows you to calculate average values based on specific conditions. 
For example, you may want to calculate the average price of products within a certain category or for a specific customer segment. 
To achieve this, you can use the FILTER function in combination with AVERAGEX.
AVERAGE: Returns the average (arithmetic mean) of all the numbers in a column. 
AVERAGEA: Returns the average (arithmetic mean) of the values in a column. 
AVERAGEX: Calculates the average (arithmetic mean) of a set of expressions evaluated over a table.
The SWITCH function in DAX is a powerful and versatile tool for financial modeling in Power BI.
Its ability to handle multiple conditions and return different results based on those conditions makes it an invaluable asset for dynamic calculations, 
simplified logic, and conditional statement.

Here I have used Average function and Switch case for finding the measure of inome of per person.